Server cleanup, split out loading handler
